# fix: cli performance