+ faas-cli动态加载插件，相关性能对比
from:
```
Benchmark:
  faas-cli invoke        x 0.40 ops/sec ±3.09% (5 runs sampled)
  faas-cli help          x 0.54 ops/sec ±5.14% (6 runs sampled)
  faas-cli test          x 0.48 ops/sec ±0.58% (6 runs sampled)
```
to:
```
Benchmark:
  faas-cli invoke        x 0.57 ops/sec ±0.71% (6 runs sampled)
  faas-cli help          x 0.54 ops/sec ±0.64% (6 runs sampled)
  faas-cli test          x 0.87 ops/sec ±0.54% (7 runs sampled)
```

即：
invoke 提升 42.5%
help 无提升
test 提升 81.25%